feat: 아티클 타입 정리 및 뷰 로직 수정 (#ATR-31)
