+++ apps/service/app/layout.tsx
@@ -19,7 +19,9 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body className="bg-gray-50 text-gray-700 dark:bg-gray-900 dark:text-gray-100">
+        {children}
+      </body>
     </html>
   )
 }
