+++ apps/service/entities/article/types/newsletter.ts
@@ -1,5 +1,14 @@
-import type { NewsletterCategory } from './newsletter-category'
-import type { UploadDays } from './upload-days'
+import { NEWSLETTER_CATEGORY } from '../constants'
+import { UPLOAD_DAYS } from '../constants'
+
+export type UploadDays = keyof typeof UPLOAD_DAYS
+
+export type ValueOfUploadDays = (typeof UPLOAD_DAYS)[UploadDays]
+
+export type NewsletterCategory = keyof typeof NEWSLETTER_CATEGORY
+
+export type ValueOfNewsletterCategory =
+  (typeof NEWSLETTER_CATEGORY)[NewsletterCategory]
 
 export interface Newsletter {
   id: string | number
